Added files for the medium post
The Python notebook and the Power BI desktop file are used to build a pipeline and embed it in Power BI. To make it work, you will have to run the code in the notebook. In the Power BI file you have to alter the path in the Python visualization that points to the serialized model.